bit of refractoring
